moves SciStarter API call to a util file
